refactor(l5): move extraction from src/ into server/domain behind an L5 adapter (issue 864)
Moved (issue 864, L5 of _plans/server-hardening-ladder.md):

  src/extraction.ts -> server/domain/extraction.ts

The moved module meets the server/ standard whole. The one change it needed
was backgroundExtract, which took six positional parameters and so failed the
four-parameter signature rule. The four per-row trailing parameters are now
one named object:

  export function backgroundExtract(
    pool: pg.Pool,
    table: string,
    target: BackgroundExtractTarget,
  ): void

BackgroundExtractTarget carries entryId, namespace, text, and the optional
existingTags. The split follows the six call sites: every one of them passes
a pool and a table name it chose, then four fields that all describe the same
single row. Every other export keeps its signature untouched.

src/extraction.ts is now the L5 adapter (M9): 29 code lines, importing only
from server/, re-exporting the unchanged exports with `export *` and defining
backgroundExtract with the old six-positional form over the new one. Every
legacy src/ caller (src/rest-api.ts, src/tools/log-thought.ts,
src/tools/log-decision.ts, src/drop-folder-collector.ts,
scripts/bulk-import.ts) and both test files are untouched and pass through it.

server/maintenance/graph-derivation-handler.ts was the one server/ importer
and is repointed to ../domain/extraction.ts per M2.

Tests left in src/ (retire or move at L6), per M1 EXCEPTION:
  src/__tests__/extraction.test.ts
  src/db/migrations/027_source_registry.test.ts

Runtime closure of src/ under server/main.ts: 35 -> 34.

Verified:
  bunx tsc --noEmit -> exit 0
  oxlint --deny-warnings --config .oxlintrc.json on all three staged paths -> exit 0
  bun run test:isolated src/__tests__/extraction.test.ts src/db/migrations/027_source_registry.test.ts -> 31 pass, 0 fail
  bash scripts/done-means/864-moved-out-of-src.sh src/extraction.ts -> PASS
